feat: refactor DashBoard styles
