--- src/containers/DashBoard/components/UsageStatus/styles.ts
@@ -1,23 +1,0 @@
-import { makeStyles, Theme, createStyles } from '@material-ui/core/styles'
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    usageStatusContainer: {
-      display: 'grid',
-      gridTemplateColumns: 'repeat(2, 1fr)',
-      gridColumnGap: 24,
-      height: 375,
-    },
-
-    paper: {
-      padding: 16,
-    },
-
-    cpuWrapper: {
-      gridTemplateColumns: '1fr',
-      marginTop: 62,
-    },
-  }),
-)
-
-export default useStyles
